emergency fix to ffmpeg command that just started breaking for no reason???

+++ Windows/MainWindow/Util/AudioRecordingUtils.cs
@@ -1,4 +1,5 @@
 ﻿using AudioReplacer.Util;
+using AudioReplacer.Util.Logger;
 using System;
 using System.IO;
 using System.Threading.Tasks;
@@ -6,26 +7,18 @@ using Windows.Media;
 using Windows.Media.Capture;
 using Windows.Media.MediaProperties;
 using Windows.Storage;
-using AudioReplacer.Util.Logger;
 
 namespace AudioReplacer.Windows.MainWindow.Util;
 public class AudioRecordingUtils
 {
-    private float pitchChange = 1;
-    private string effectCommand = string.Empty;
+    public float pitchChange = 1;
+    public string effectCommand = "";
     public bool requiresExtraEdits;
     private MediaCapture recordingCapture;
 
     public AudioRecordingUtils()
     {
         Task.Run(InitializeMediaCapture);
-    }
-
-    public void SetEffectCommands(float newPitch = 1, string newEffect = "", bool extraEdits = false)
-    {
-        pitchChange = newPitch;
-        effectCommand = newEffect;
-        requiresExtraEdits = extraEdits;
     }
 
     private async Task InitializeMediaCapture()
@@ -56,31 +49,33 @@ public class AudioRecordingUtils
     public async Task StopRecordingAudio(bool discarding = false)
     {
         var file = ProjectFileUtils.GetOutFilePath();
-        switch (discarding)
-        {
-            case true:
-                await recordingCapture.StopRecordAsync();
-                File.Delete(file);
-                return;
-            case false:
-                await Task.Delay(App.AppSettings.RecordEndWaitTime);
-                await recordingCapture.StopRecordAsync();
-                await ApplyFilters(file);
-                return;
-        }
+        await Task.Delay(App.AppSettings.RecordEndWaitTime);
+        await recordingCapture.StopRecordAsync();
+        await ApplyFilters(file);
+    }
+
+    [Log]
+    public async Task CancelRecording()
+    {
+        await recordingCapture.StopRecordAsync();
+        File.Delete(ProjectFileUtils.GetOutFilePath());
     }
 
     [Log]
     private async Task ApplyFilters(string file)
     {
-        var outFile = $"{file}0.wav"; // Temporary name. FFmpeg doesn't like it when the user tries to set the output as the input since
-        var validatedPitchChange = MathF.Max(pitchChange, 0.001f); // Pitch values below zero do not work
-        var command = string.IsNullOrEmpty(effectCommand)
+        var tempOutFile = $"{file}.wav";
+        float validatedPitchChange = MathF.Max(pitchChange, 0.001f);
+
+        string filter = string.IsNullOrWhiteSpace(effectCommand)
             ? $"rubberband=pitch={validatedPitchChange}"
             : $"rubberband=pitch={validatedPitchChange}, {effectCommand}";
-        await File.WriteAllTextAsync(Path.Join(Generic.ExtraApplicationData, "coolThing.txt"), command);
-        await Generic.SpawnProcess($"{Generic.FfmpegPath}", $"-i {file} -af \"{command}\" -y {outFile}");
-        File.Delete(file);
-        File.Move(outFile, file);
+        await Generic.SpawnProcess(Generic.FfmpegPath, $"-i \"{file}\" -af \"{filter}\" -y {tempOutFile}");
+
+        if (File.Exists(tempOutFile))
+        {
+            File.Delete(file);
+            File.Move(tempOutFile, file);
+        }
     }
 }
